rename +~-~*~-tests.js to arithm-tests.js. Fixes #64.
