+++ test/browser/dsp/index.js
@@ -5,5 +5,5 @@ require('./line~-tests.js')
 require('./sig~-tests.js')
 require('./tabread~-tests.js')
 require('./lop~hip~bp~vcf~-tests.js')
-require('./+~-~*~-tests.js')
+require('./arithm-tests.js')
 require('./delwrite~delread~-tests.js')
